restructure into separate files for each audio recorder
