+++ src/acoupi/components/audio_recorder/__init__.py
@@ -1,0 +1,13 @@
+from acoupi.components.audio_recorder.microphone_config import MicrophoneConfig
+from acoupi.components.audio_recorder.pipewire_recorder import PWRecorder
+from acoupi.components.audio_recorder.pyaudio_recorder import (
+    TMP_PATH,
+    PyAudioRecorder,
+)
+
+__all__ = [
+    "MicrophoneConfig",
+    "PWRecorder",
+    "PyAudioRecorder",
+    "TMP_PATH",
+]
